Move abstraction into new package.
